requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
idna==3.3
Paginator==0.5.1
Pillow==9.1.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
